// src/types/air-quality-data.ts
export interface AirQualityData {
  coFlag: string | null;
  coGrade: string;
  coValue: string;
  dataTime: string;
  khaiGrade: string;
  khaiValue: string;
  no2Flag: string | null;
  no2Grade: string;
  no2Value: string;
  o3Flag: string | null;
  o3Grade: string;
  o3Value: string;
  pm10Flag: string | null;
  pm10Grade: string;
  pm10Value: string;
  pm25Flag: string | null;
  pm25Grade: string;
  pm25Value: string;
  sidoName: string;
  so2Flag: string | null;
  so2Grade: string;
  so2Value: string;
  stationName: string;
}